feat(community,reports,gems): add v2 pages with BFF proxy and feature flag routing
Extract legacy pages and create v2 variants using patchAdminAxios for BFF proxy.
Community v2 removes useAuth dependency. Server component routers gate on
isAdminShellV2Enabled + getRouteMode.
